# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/search": {
      "get": {
        "operationId": "UsersController_search",
        "summary": "등록 사용자 검색 (닉네임)",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "커서(seq 또는 id 또는 ISO 타임스탬프)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": true,
            "in": "query",
            "description": "검색 키워드(닉네임)",
            "schema": {
              "minLength": 1,
              "maxLength": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/users/{userId}/profile": {
      "get": {
        "operationId": "UsersController_getProfile",
        "summary": "사용자 프로필 조회 (같은 룸/연락처만)",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/profile": {
      "post": {
        "operationId": "ProfileController_create",
        "summary": "최초 온보딩 (닉네임/아바타/상태메시지)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ProfileController_update",
        "summary": "프로필 부분 수정",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "tags": [
          "profile"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms": {
      "get": {
        "operationId": "RoomsController_list",
        "summary": "내 채팅방 목록 (최근 활동순, 미읽음 수 포함)",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "커서(seq 또는 id 또는 ISO 타임스탬프)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지 크기",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRoomsResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "summary": "방 생성 (1:1 중복 시 기존 방 반환 / 그룹)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}": {
      "get": {
        "operationId": "RoomsController_detail",
        "summary": "방 상세 + 멤버 목록",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "summary": "방 정보 수정 (그룹, OWNER/ADMIN)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}/members": {
      "post": {
        "operationId": "RoomsController_addMembers",
        "summary": "멤버 추가 (그룹, OWNER/ADMIN)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMembersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/rooms/{roomId}/members/{userId}": {
      "patch": {
        "operationId": "RoomsController_updateMemberRole",
        "summary": "멤버 역할 변경 (그룹, OWNER)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomViewResponse"
                }
              }
            }
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RoomsController_removeMember",
        "summary": "멤버 탈퇴(본인) 또는 추방(OWNER/ADMIN)",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "탈퇴/추방 완료(본문 없음)"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/push/tokens": {
      "post": {
        "operationId": "PushController_register",
        "summary": "FCM 토큰 등록",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTokenResponse"
                }
              }
            }
          }
        },
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/push/tokens/{tokenId}": {
      "delete": {
        "operationId": "PushController_remove",
        "summary": "FCM 토큰 삭제",
        "parameters": [
          {
            "name": "tokenId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료(본문 없음)"
          }
        },
        "tags": [
          "push"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "인택톡(Intaktok) API",
    "description": "폐쇄형 실시간 채팅 플랫폼 백엔드 API (v2.0, E2EE 제외)",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PublicUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "nickname",
          "avatarUrl"
        ]
      },
      "PaginatedUsersResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "다음 페이지 커서. null이면 마지막 페이지입니다."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUserResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "items"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "statusMessage": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "lastSeenAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "마지막 접속 시각"
          }
        },
        "required": [
          "id",
          "nickname",
          "statusMessage",
          "avatarUrl",
          "lastSeenAt"
        ]
      },
      "CreateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30,
            "example": "택이"
          },
          "statusMessage": {
            "type": "string",
            "maxLength": 200,
            "example": "안녕하세요"
          },
          "avatarMediaId": {
            "type": "string",
            "description": "아바타 미디어 id"
          }
        },
        "required": [
          "nickname"
        ]
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "statusMessage": {
            "type": "string",
            "nullable": true
          },
          "avatarMediaId": {
            "type": "string",
            "nullable": true
          },
          "onboardedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "null이면 온보딩 미완료"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "표시용 아바타 URL(없으면 null)"
          }
        },
        "required": [
          "userId",
          "nickname",
          "statusMessage",
          "avatarMediaId",
          "onboardedAt",
          "updatedAt",
          "avatarUrl"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30
          },
          "statusMessage": {
            "type": "string",
            "maxLength": 200
          },
          "avatarMediaId": {
            "type": "string"
          }
        }
      },
      "RoomMemberResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "MEMBER"
            ]
          }
        },
        "required": [
          "userId",
          "nickname",
          "avatarUrl",
          "role"
        ]
      },
      "RoomLastMessageResponse": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "number",
            "description": "룸 내 단조 증가 시퀀스"
          },
          "type": {
            "type": "string",
            "enum": [
              "TEXT",
              "IMAGE",
              "GIF",
              "VIDEO",
              "FILE",
              "SYSTEM"
            ]
          },
          "preview": {
            "type": "string",
            "description": "미리보기 텍스트(삭제 시 placeholder)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "seq",
          "type",
          "preview",
          "createdAt"
        ]
      },
      "RoomViewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "DIRECT",
              "GROUP"
            ]
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "GROUP 방 이름"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomMemberResponse"
            }
          },
          "lastMessage": {
            "nullable": true,
            "description": "마지막 메시지(없으면 null)",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoomLastMessageResponse"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "미읽음 메시지 수"
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "avatarUrl",
          "createdBy",
          "createdAt",
          "lastMessageAt",
          "members"
        ]
      },
      "PaginatedRoomsResponse": {
        "type": "object",
        "properties": {
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "다음 페이지 커서. null이면 마지막 페이지입니다."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomViewResponse"
            }
          }
        },
        "required": [
          "nextCursor",
          "items"
        ]
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DIRECT",
              "GROUP"
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 80,
            "description": "GROUP 방 이름"
          },
          "avatarMediaId": {
            "type": "string"
          },
          "memberUserIds": {
            "description": "상대(DIRECT=1명) 또는 초대 멤버(GROUP)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "memberUserIds"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80
          },
          "avatarMediaId": {
            "type": "string"
          }
        }
      },
      "AddMembersDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userIds"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MEMBER",
              "OWNER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
      "RegisterTokenDto": {
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string",
            "maxLength": 4096
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          },
          "deviceId": {
            "type": "string",
            "maxLength": 128
          }
        },
        "required": [
          "fcmToken",
          "platform"
        ]
      },
      "PushTokenResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "enum": [
              "ios",
              "android",
              "web"
            ]
          }
        },
        "required": [
          "id",
          "platform"
        ]
      },
...
    }
  }
}